Write a descriptive README
I came to this project through Octokit, and found the previous README to
be somewhat lacking in terms of describing what this library actually
does.

So, I poked around in the Octokit and Sawyer internals enough to get an
understanding of what the project is and how it works, and then I wrote a
README based on what I'd learned.

This fixes #11, and fixes #37.