Add `dig` and `fetch` to `Sawyer::Resource`
These convenience methods make it easier to work with `Resource`s as a
consumer, especially when they come from APIs that have a lot of
nesting.

Supporting these methods, in addition to `[]`, `[]=`, `each`, and all
the other enumerable methods, allows consumers to treat a
`Sawyer::Resource` effectively as a Hash for the attributes without
having to convert it to one and lose the extra behavior that the
`Resource` provides.